Adjusted enum type * Adjusted pet catch enum type to match player storage type. Thanks to @Lemongrass3110!

+++ src/map/pet.hpp
@@ -47,7 +47,7 @@ extern struct s_pet_db pet_db[MAX_PET_DB];
 
 enum e_pet_itemtype : uint8 { PET_CLASS,PET_CATCH,PET_EGG,PET_EQUIP,PET_FOOD };
 
-enum e_pet_catch : uint8 {
+enum e_pet_catch : uint16 {
 	PET_CATCH_FAIL = 0, ///< A catch attempt failed
 	PET_CATCH_UNIVERSAL = 1, ///< The catch attempt is universal (ignoring MD_STATUS_IMMUNE/Boss)
 };
